feat(support-bundle): move SplitTroubleshootSecretLabelSelector to specs.go
